refactor: Update BlobUploadService interface to include container name parameter
This commit updates the `BlobUploadService` interface in the `blob_upload_service.py` file to include a `container_name` parameter in the `upload_blob` method. This change allows for specifying the name of the container where the blob will be uploaded. It improves the flexibility and configurability of the codebase by providing a way to upload blobs to different containers within the storage service.