refactor: Update StorageBlobUploadService to support uploading blobs to different containers
This commit updates the `StorageBlobUploadService` class in the `storage_blob_upload_service.py` file to support uploading blobs to different containers in Azure Blob Storage. It adds a new parameter `container_name` to the `upload_blob` method, allowing the user to specify the name of the container where the blob will be uploaded. This change improves the flexibility and configurability of the codebase, enabling the uploading of blobs to specific containers based on requirements.